+++ table/block_based/block_based_table_factory.h
@@ -13,6 +13,7 @@
 #include <memory>
 #include <string>
 
+#include "memory/cache_capacity_based_memory_allocator.h"
 #include "port/port.h"
 #include "rocksdb/flush_block_policy.h"
 #include "rocksdb/table.h"
@@ -89,6 +90,8 @@ class BlockBasedTableFactory : public TableFactory {
 
  private:
   BlockBasedTableOptions table_options_;
+  std::shared_ptr<CacheCapacityBasedMemoryAllocator>
+      table_reader_mem_allocator_;
   mutable TailPrefetchStats tail_prefetch_stats_;
 };