+++ table/block_based/block_based_table_reader_test.cc
@@ -5,6 +5,7 @@
 
 #include "table/block_based/block_based_table_reader.h"
 
+#include "db/db_test_util.h"
 #include "db/table_properties_collector.h"
 #include "file/file_util.h"
 #include "options/options_helper.h"
@@ -247,6 +248,259 @@ TEST_P(BlockBasedTableReaderTest, MultiGet) {
     ASSERT_TRUE(table->TEST_KeyInCache(ReadOptions(), keys[i]));
     ASSERT_EQ(values[i].ToString(), kv[keys[i].ToString()]);
   }
+}
+
+class BlockBasedTableReaderResOnlyCache : public CacheWrapper {
+ public:
+  explicit BlockBasedTableReaderResOnlyCache(std::shared_ptr<Cache> target)
+      : CacheWrapper(std::move(target)) {}
+
+  using Cache::Insert;
+  Status Insert(const Slice& key, void* value, size_t charge,
+                void (*deleter)(const Slice& key, void* value),
+                Handle** handle = nullptr,
+                Priority priority = Priority::LOW) override {
+    if (deleter == kNoopDeleterForBlockBasedTableReader) {
+      return target_->Insert(key, value, charge, deleter, handle, priority);
+    } else {
+      return Status::OK();
+    }
+  }
+
+  using Cache::Release;
+  bool Release(Handle* handle, bool force_erase = false) override {
+    auto deleter = GetDeleter(handle);
+    if (deleter == kNoopDeleterForBlockBasedTableReader) {
+      return target_->Release(handle, force_erase);
+    } else {
+      return true;
+    }
+  }
+
+ private:
+  static const Cache::DeleterFn kNoopDeleterForBlockBasedTableReader;
+};
+
+const Cache::DeleterFn
+    BlockBasedTableReaderResOnlyCache::kNoopDeleterForBlockBasedTableReader =
+        CacheReservationManager::TEST_GetNoopDeleterForRole<
+            CacheEntryRole::kBlockBasedTableReader>();
+
+class BlockBasedTableReaderCapMemoryTest
+    : public testing::Test,
+      public testing::WithParamInterface<
+          bool /* reserve_table_builder_memory */> {
+ protected:
+  void SetUp() override {
+    env_ = Env::Default();
+    fs_ = FileSystem::Default();
+    test_dir_ = test::PerThreadDBPath("block_based_table_reader_test");
+    ASSERT_OK(fs_->CreateDir(test_dir_, IOOptions(), nullptr));
+
+    reserve_table_reader_memory_ = GetParam();
+  }
+
+  void TearDown() override { EXPECT_OK(DestroyDir(env_, test_dir_)); }
+
+  void ConfugreOptions(std::size_t reserve_table_reader_memory,
+                       std::size_t cache_capacity = 1024) {
+    BlockBasedTableOptions table_options;
+
+    table_reader_res_only_cache_.reset(new BlockBasedTableReaderResOnlyCache(
+        NewLRUCache(cache_capacity, 0 /* num_shard_bits */,
+                    true /* strict_capacity_limit */)));
+    table_options.block_cache = table_reader_res_only_cache_;
+
+    table_options.reserve_table_reader_memory = reserve_table_reader_memory;
+
+    table_options.filter_policy.reset(NewBloomFilterPolicy(10, false));
+    table_options.partition_filters = true;
+    table_options.index_type = BlockBasedTableOptions::kTwoLevelIndexSearch;
+
+    options_.table_factory.reset(NewBlockBasedTableFactory(table_options));
+  }
+
+  void CreateTable(const std::string& table_name) {
+    std::unique_ptr<WritableFileWriter> writer;
+    NewFileWriter(table_name, &writer);
+
+    ImmutableOptions ioptions(options_);
+    InternalKeyComparator comparator(options_.comparator);
+    ColumnFamilyOptions cf_options;
+    MutableCFOptions moptions(cf_options);
+    IntTblPropCollectorFactories factories;
+    std::unique_ptr<TableBuilder> table_builder(
+        options_.table_factory->NewTableBuilder(
+            TableBuilderOptions(ioptions, moptions, comparator, &factories,
+                                CompressionType::kNoCompression,
+                                CompressionOptions(), 0 /* column_family_id */,
+                                kDefaultColumnFamilyName, -1 /* level */),
+            writer.get()));
+
+    Random rnd(101);
+    for (uint32_t i = 0; i < 100; ++i) {
+      char k[9] = {0};
+      // Internal key is constructed directly from this key,
+      // and internal key size is required to be >= 8 bytes,
+      // so use %08u as the format string.
+      sprintf(k, "%08u", i);
+      std::string key = ToInternalKey(k);
+      std::string val = rnd.RandomString(256);
+      table_builder->Add(key, val);
+    }
+
+    ASSERT_OK(table_builder->Finish());
+  }
+
+  void NewBlockBasedTableReader(const std::string& table_name,
+                                std::unique_ptr<TableReader>* table_reader,
+                                Status* create_table_reader_status) {
+    assert(create_table_reader_status);
+
+    ReadOptions ro;
+
+    const ImmutableOptions ioptions(options_);
+    const MutableCFOptions moptions(options_);
+    const InternalKeyComparator ikc(options_.comparator);
+    TableReaderOptions table_reader_options = TableReaderOptions(
+        ioptions, moptions.prefix_extractor, EnvOptions(), ikc);
+
+    std::unique_ptr<RandomAccessFileReader> file_reader;
+    const FileOptions foptions;
+    NewFileReader(table_name, foptions, &file_reader);
+    uint64_t file_size = 0;
+    ASSERT_OK(env_->GetFileSize(Path(table_name), &file_size));
+
+    (*create_table_reader_status) = options_.table_factory->NewTableReader(
+        table_reader_options, std::move(file_reader), file_size, table_reader,
+        false);
+  }
+
+  std::size_t ApproximateTableReaderMem() {
+    ConfugreOptions(false /* reserve_table_reader_mem */);
+
+    std::size_t approx_table_reader_mem = 0;
+    std::string table_name = "table_for_approx_table_reader_mem";
+    CreateTable(table_name);
+    std::unique_ptr<TableReader> table_reader;
+    Status s;
+    NewBlockBasedTableReader(table_name, &table_reader, &s);
+    assert(s.ok());
+    approx_table_reader_mem = table_reader->ApproximateMemoryUsage();
+    assert(approx_table_reader_mem > 0);
+    return approx_table_reader_mem;
+  }
+
+  static std::size_t ApproximateMaxTableReaderNumBeforeCacheFull(
+      std::size_t cache_capacity, std::size_t approx_table_reader_mem) {
+    // To make calculation easier for testing
+    assert(cache_capacity % CacheReservationManager::GetDummyEntrySize() == 0);
+
+    // We need to subtract 1 for max_num_dummy_entry to account for dummy
+    // entry's overhead
+    std::size_t max_num_dummy_entry =
+        (size_t)std::floor((1.0 * cache_capacity /
+                            CacheReservationManager::GetDummyEntrySize())) -
+        1;
+    std::size_t cache_capacity_rounded_to_dummy_entry_multiples =
+        max_num_dummy_entry * CacheReservationManager::GetDummyEntrySize();
+    std::size_t max_table_reader_num = static_cast<std::size_t>(
+        std::floor(1.0 * cache_capacity_rounded_to_dummy_entry_multiples /
+                   approx_table_reader_mem));
+
+    return max_table_reader_num;
+  }
+
+  bool reserve_table_reader_memory_;
+  std::shared_ptr<BlockBasedTableReaderResOnlyCache>
+      table_reader_res_only_cache_;
+
+ private:
+  void NewFileWriter(const std::string& filename,
+                     std::unique_ptr<WritableFileWriter>* writer) {
+    std::string path = Path(filename);
+    EnvOptions env_options;
+    FileOptions foptions;
+    std::unique_ptr<FSWritableFile> file;
+    ASSERT_OK(fs_->NewWritableFile(path, foptions, &file, nullptr));
+    writer->reset(new WritableFileWriter(std::move(file), path, env_options));
+  }
+
+  void NewFileReader(const std::string& filename, const FileOptions& opt,
+                     std::unique_ptr<RandomAccessFileReader>* reader) {
+    std::string path = Path(filename);
+    std::unique_ptr<FSRandomAccessFile> f;
+    ASSERT_OK(fs_->NewRandomAccessFile(path, opt, &f, nullptr));
+    reader->reset(new RandomAccessFileReader(std::move(f), path,
+                                             env_->GetSystemClock().get()));
+  }
+
+  std::string Path(const std::string& fname) { return test_dir_ + "/" + fname; }
+
+  std::string ToInternalKey(const std::string& key) {
+    InternalKey internal_key(key, 0, ValueType::kTypeValue);
+    return internal_key.Encode().ToString();
+  }
+
+  Env* env_;
+  std::shared_ptr<FileSystem> fs_;
+  std::string test_dir_;
+  Options options_;
+};
+
+INSTANTIATE_TEST_CASE_P(CapMemoryUsageUnderCacheCapacity,
+                        BlockBasedTableReaderCapMemoryTest, ::testing::Bool());
+
+TEST_P(BlockBasedTableReaderCapMemoryTest, CapMemoryUsageUnderCacheCapacity) {
+  std::size_t cache_capacity =
+      10 * CacheReservationManager::GetDummyEntrySize();
+  std::size_t mem_per_table_reader = ApproximateTableReaderMem();
+  std::size_t max_table_reader_num = BlockBasedTableReaderCapMemoryTest::
+      ApproximateMaxTableReaderNumBeforeCacheFull(cache_capacity,
+                                                  mem_per_table_reader);
+
+  ConfugreOptions(true, cache_capacity);
+  std::vector<std::unique_ptr<TableReader>> table_reader_objects;
+  Status s = Status::OK();
+  std::size_t opened_table_reader_num = 0;
+  std::string table_name;
+
+  // Keep creating BlockBasedTableReader till hiting the memory limit based on
+  // cache capacity and creation fails
+  while (s.ok()) {
+    table_name = "table_" + std::to_string(opened_table_reader_num);
+    CreateTable(table_name);
+    table_reader_objects.push_back(std::unique_ptr<TableReader>());
+    NewBlockBasedTableReader(table_name, &(table_reader_objects.back()), &s);
+    if (s.ok()) {
+      ++opened_table_reader_num;
+    }
+  }
+  EXPECT_TRUE(s.IsMemoryLimit());
+  EXPECT_TRUE(s.ToString().find("memory limit based on cache capacity") !=
+              std::string::npos);
+  EXPECT_GE(opened_table_reader_num, max_table_reader_num * 9 / 10);
+  EXPECT_LE(opened_table_reader_num, max_table_reader_num);
+
+  std::size_t updated_max_table_reader_num =
+      BlockBasedTableReaderCapMemoryTest::
+          ApproximateMaxTableReaderNumBeforeCacheFull(cache_capacity / 2,
+                                                      mem_per_table_reader);
+
+  // Keep deleting BlockBasedTableReader to lower down memory usage from limit
+  // so that next creation succeeds
+  while (opened_table_reader_num >= updated_max_table_reader_num) {
+    table_reader_objects.pop_back();
+    --opened_table_reader_num;
+  }
+  table_name = "table_for_successful_table_reader_open";
+  CreateTable(table_name);
+  table_reader_objects.push_back(std::unique_ptr<TableReader>());
+  NewBlockBasedTableReader(table_name, &(table_reader_objects.back()), &s);
+  EXPECT_TRUE(s.ok()) << s.ToString();
+
+  table_reader_objects.clear();
+  EXPECT_EQ(table_reader_res_only_cache_->GetPinnedUsage(), 0);
 }
 
 class BlockBasedTableReaderTestVerifyChecksum